+++ src/Support/Changelog.md
@@ -1,5 +1,14 @@
 # Vanilo Support Module Changelog
 
+## 5.x Series
+
+## Unreleased
+##### 2025-XX-XX
+
+- Dropped PHP 8.2 Support
+- Added Laravel 12 Support
+- PHP 8.4 deprecation fixes
+
 ## 4.x Series
 
 ## 4.2.0